iOS: Add minimal file access
We can use devicectl to access some parts of the file system that fall
into different domains, like crash reports and the developer's
application directories.

Add a file access that provides a virtual root directory with these
domains as the "subdirectories", and access to the file systems below
them.

There are some restrictions: We cannot rename or remove files with the
devicectl API. So we need to explicitly disable atomic save, and also
the editor manager's auto-save to avoid accumulating .autosave files.
Add another supportsRemovingFiles hook in the device file access for the
latter (the former was already added in a preceeding patch).

+++ src/plugins/coreplugin/editormanager/editormanager.cpp
@@ -876,7 +876,8 @@ IEditor *EditorManagerPrivate::openEditor(EditorView *view, const FilePath &file
 
     FilePath realFp = autoSaveName(filePath);
     if (!filePath.exists() || !realFp.exists() || filePath.lastModified() >= realFp.lastModified()) {
-        realFp.removeFile();
+        if (realFp.supportsRemoving())
+            realFp.removeFile();
         realFp = filePath;
     }
 
@@ -2603,8 +2604,8 @@ void EditorManagerPrivate::autoSave()
             continue;
         const FilePath saveName = autoSaveName(document->filePath());
         const FilePath savePath = saveName.absolutePath();
-        if (document->filePath().isEmpty()
-                || !savePath.isWritableDir()) // FIXME: save them to a dedicated directory
+        if (document->filePath().isEmpty() || !document->filePath().supportsRemoving()
+            || !savePath.isWritableDir()) // FIXME: save them to a dedicated directory
             continue;
         if (Result<> res = document->autoSave(saveName); !res)
             errors << res.error();
